jsonl: consolidate cutover tests and pin v8/v9 fixture shape
Merges cutover_v4_test.go, cutover_v7_test.go, and cutover_v9_test.go
into cutover_test.go so every cutover-related test lives in one file.

Adds assertTableShape and applies it (via assertV8V9Shape) to both
seedV8DBWithOrphans and seedV9SchemaDB. The fixtures previously
bootstrapped the current schema and dropped "the v10 additions" by
hand; nothing verified the drop list was complete. That's the gap
that let "no such column: metadata" reach real v9 cutover without
any test failing. With the guard in place, a future v11 column added
without updating dropV10Additions fails this assertion instead.

The seedV8DBWithOrphans fixture now actually models real v8 shape
(no metadata, no revision, no recurrence linkage), so the v8 cutover
tests exercise the same export branches a real v8 source would hit.
